Modified settings.py to use local_settings.py for personal settings
All the local deployment settings, like DATABASES or STATIC_ROOT or
SECRET_KEY or other should be kept in a separate file named
local_settings.py which should not be added to the repository.
(It's already added to .gitignore)
The file should reside in the same directory as the settings.py

This way it's easier to tailor custom deployments - be it developer local
environment or an acceptance server or even a production server.
It gives flexibility in accessing the settings' values as well, be it
hard coded or from environment variables - whatever suits your needs.

Also included is a small typo fix of the manage.py script.