Fix UserSerializer create method
In general *create* method should just return a created object instance. In UserSerializer create was returning the password alongside user object. The password was needed for the mail to the registered user.
The password should not be sent plain-text in any communication (especially in mail). This has been changed and no password is being send in the mail.
Of course this requires changing the way a new user can authenticate himeslf for the first time.

Along the way there was a fix in send_to_user_created_by_hr method. Request object should not be send as a parameter to any method, because that make this method untestable.